dating-mentor-flask/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
Werkzeug==2.3.7

# Image processing and OCR
Pillow==10.0.0
pytesseract==0.3.10

# AI and NLP
langchain==0.0.340
langchain-community==0.0.10
gigachat==0.1.17
nltk==3.8.1
language-tool-python==2.7.1

# Utilities
python-dotenv==1.0.0
uuid==1.30

# Production server (optional)
gunicorn==21.2.0
